+++ sdk/ai/Azure.AI.Extensions.OpenAI/tests/Samples/Sample_OpenAPI.cs
@@ -42,7 +42,7 @@ public class Sample_OpenAPI : ProjectsOpenAITestBase
         #endregion
         #region Snippet:Sample_CreateAgent_OpenAPI_Async
         string filePath = GetFile();
-        OpenAPIFunctionDefinition toolDefinition = new(
+        OpenApiFunctionDefinition toolDefinition = new(
             name: "get_weather",
             specificationBytes: BinaryData.FromBytes(File.ReadAllBytes(filePath)),
             authentication: new OpenAPIAnonymousAuthenticationDetails()
@@ -87,7 +87,7 @@ public class Sample_OpenAPI : ProjectsOpenAITestBase
         AIProjectClient projectClient = new(endpoint: new Uri(projectEndpoint), tokenProvider: new DefaultAzureCredential());
         #region Snippet:Sample_CreateAgent_OpenAPI_Sync
         string filePath = GetFile();
-        OpenAPIFunctionDefinition toolDefinition = new(
+        OpenApiFunctionDefinition toolDefinition = new(
             name: "get_weather",
             specificationBytes: BinaryData.FromBytes(File.ReadAllBytes(filePath)),
             authentication: new OpenAPIAnonymousAuthenticationDetails()