+++ sdk/ai/Azure.AI.Projects.Agents/tests/Samples/Sample_HostedAgent.cs
@@ -36,7 +36,7 @@ public class Sample_HostedAgent : SamplesBase
     private static  HostedAgentDefinition GetAgentDefinition(string dockerImage, string modelDeploymentName, string accountId, string applicationInsightConnectionString, string projectEndpoint)
     {
         HostedAgentDefinition agentDefinition = new(
-            containerProtocolVersions: [new ProtocolVersionRecord(AgentCommunicationMethod.ActivityProtocol, "v1")],
+            versions: [new ProtocolVersionRecord(AgentProtocol.ActivityProtocol, "v1")],
             cpu: "1",
             memory: "2Gi"
         )