+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Other Changes
 
+- Updated `Microsoft.Identity.Client` and `Microsoft.Identity.Client.Extensions.Msal` dependencies to version 4.83.1.
+
 ## 1.18.0 (2026-02-25)
 
 ### Features Added